Fixes for unit test failures
The first defect was that the source handle output argument was not being initialized in add_source. The second defect was that SACN_DMX_MERGER_SOURCE_INVALID was defined to a negative value when it is supposed to be used with unsigned short handles - so it was converted to the uint16_t version of "-1".