SACN-188 - The library is not unsubscribing sockets on universe removal/change
This isn't just adding an unsubscribe call in a function. I found that the code needed to be refactored so that the subscribing and unsubscribing happens on the main thread, since 99% of the time the socket already belongs to the thread and could be getting polled. So I added a couple more queues to make that work.